converted C# implementation to .NET Core RC2
