+++ csharp/NuGet.Config
@@ -1,0 +1,9 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+  <activePackageSource>
+    <add key="dev_Rendler" value="dev_nuget_feed" />
+  </activePackageSource>
+  <packageSources>
+    <add key="dev_Rendler" value="dev_nuget_feed" />
+  </packageSources>
+</configuration>
