--- csharp/Rendler/Executors/Messages/CrawlResultMessage.cs
@@ -1,14 +1,0 @@
-﻿using System.Runtime.Serialization;
-
-namespace Rendler.Executors.Messages
-{
-    [DataContract]
-    public class CrawlResultMessage
-    {
-        [DataMember]
-        public string Url { get; set; }
-
-        [DataMember]
-        public string[] Links { get; set; }
-    }
-}
