--- csharp/Rendler/Executors/Messages/Message.cs
@@ -1,14 +1,0 @@
-﻿using System.Runtime.Serialization;
-
-namespace Rendler.Executors.Messages
-{
-    [DataContract]
-    internal class Message
-    {
-        [DataMember]
-        public string Type { get; set; }
-
-        [DataMember]
-        public byte[] Body { get; set; }
-    }
-}
