--- csharp/Rendler/RunMode.cs
@@ -1,9 +1,0 @@
-﻿namespace Rendler
-{
-    internal enum RunMode
-    {
-        Default,
-        Scheduler,
-        Executor
-    }
-}
