+++ csharp/dev_nuget_feed/readme
@@ -1,0 +1,1 @@
+place 'mesosclr.x.x.x.nupkg' in this directory
