--- csharp/ext/readme
@@ -1,3 +1,0 @@
-directory containing the mesos-clr binaries required to build Rendler
-- mesosclr.dll
-- libmesosclr.so
