+++ csharp/src/Rendler/Executors/Messages/RenderResultMessage.cs
@@ -1,0 +1,14 @@
+﻿using System.Runtime.Serialization;
+
+namespace Rendler.Executors.Messages
+{
+    [DataContract]
+    public class RenderResultMessage
+    {
+        [DataMember]
+        public string Url { get; set; }
+
+        [DataMember]
+        public string FileName { get; set; }
+    }
+}
