+++ csharp/src/Rendler/Properties/AssemblyInfo.cs
@@ -1,0 +1,16 @@
+﻿using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("mesosclr.Rendler")]
+[assembly: AssemblyDescription("")]
+[assembly: ComVisible(false)]
+
+[assembly: AssemblyConfiguration("")]
+[assembly: AssemblyCompany("")]
+[assembly: AssemblyProduct("mesosclr.Rendler")]
+[assembly: AssemblyCopyright("")]
+[assembly: AssemblyTrademark("")]
+[assembly: AssemblyCulture("")]
+
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]
