+++ csharp/src/Rendler/RunMode.cs
@@ -1,0 +1,9 @@
+﻿namespace Rendler
+{
+    internal enum RunMode
+    {
+        Default,
+        Scheduler,
+        Executor
+    }
+}
